Small fixes as per PR review
